renamed MessageManger component file to index
